Removed `secondary_index` test
The test was meant to check the consistency of the Checksum256 encoding with the SNARKtor contracts and the WASM-compiled smart contracts in general. However, given that the `rscdk` compiler treats the encoding of Checksum256 with a different encoding compared to the C++ compiler, the test has been removed.
It could be restored once the `rscdk` issue is fixed.